requirements.txt
Django==1.11.4
certifi==2017.11.5
ipaddress==1.0.19
munge==0.4.0
mysqlclient==1.3.9
setuptools==30.2.0
tld==0.7.6
uwsgi==2.0.14
Markdown==2.6.7
bleach==2.1.3
coreapi==2.3.1
googlemaps==2.5.1

django-allauth==0.32.0
django-autocomplete-light==3.2.9
django-bootstrap3==9.0.0
django-cors-middleware==1.3.1
django-countries==4.6.1
django-crispy-forms==1.6.1
django-debug-toolbar==1.9.1
django-extensions==1.3.3
django-forms-bootstrap==3.1.0
django-grappelli==2.10.1
django-mobi==0.1.7
django-oauth-toolkit==1.0.0
django-phonenumber-field==0.6
django-reversion==2.0.9
django-rest-swagger==2.1.2
djangorestframework==3.6.3
django-tables2==1.0.4
django-vanilla-views==1.0.2
django-cors-headers==2.1.0
django-ratelimit==1.1.0

# issue #95
passlib==1.7.1
# think this is just from import?
# django admin whines if this is not installed, should re-investigate
# during #95
django-hashers-passlib==0.3

# old third_party
django-inet==0.3.2
django-handleref==0.2.0
django-namespace-perms==0.5.0
#FIXME: should change from pip once 0.6.0 is tagged and released
django-peeringdb==0.6.0
peeringdb==0.5.1
rdap==0.2.0
twentyc.rpc==0.3.5
